% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fac}
\alias{fac}
\title{Independence test of categorical variables}
\usage{
fac(n)
}
\arguments{
\item{a}{a matrix}
}
\description{
Independence test of categorical variables
}
\examples{
chisq_test(matrix(c(98,38,289,67,41,262,13,8,57,18,12,30),4,3,b=T))
chisq_test(matrix(c(341,405,105,103,11,15),2,3,b=T))
}
